refactor: Move bms part in a folder like bmson
